requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sgp4==2.23
astropy==6.0.1
numpy==1.26.4
pyproj==3.6.1
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
reportlab==4.2.2
jinja2==3.1.4
httpx==0.28.1
pytest==8.4.1
